third_party/WebKit/LayoutTests/accessibility/inline-text-bidi-bounds-for-range-expected.txt

one two אחתשתיים three four שלושהארבעה

one two אחתשתיים three four שלושהארבעה

Tests that we can compute the bounds of a range of text from the accessibility tree in bidirectional text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing bounds of word one in horizontalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word two in horizontalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word three in horizontalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word four in horizontalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word אחתשתיים in horizontalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word שלושהארבעה in horizontalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word one in verticalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word two in verticalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word three in verticalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word four in verticalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word אחתשתיים in verticalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height

Testing bounds of word שלושהארבעה in verticalParagraph
PASS axBounds.x is within 2 of rangeBounds.left
PASS axBounds.y is within 2 of rangeBounds.top
PASS axBounds.width is within 2 of rangeBounds.width
PASS axBounds.height is within 2 of rangeBounds.height
PASS successfullyParsed is true

TEST COMPLETE
